#921 worked around normalization problem

+++ NegotiatorGUI/src/negotiator/ScenarioValidator.java
@@ -1,4 +1,5 @@
 package negotiator;
+
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Map.Entry;
@@ -16,32 +17,37 @@ import negotiator.utility.EvaluatorDiscrete;
 import negotiator.utility.UtilitySpace;
 
 /**
- * Class used to validate the domain, preference profile, and the consistency between both.
+ * Class used to validate the domain, preference profile, and the consistency
+ * between both.
  * 
  * @author Mark Hendrikx
  */
 public class ScenarioValidator {
-	
+
 	private static String validateDomain(Domain domain) {
 		String errors = "";
 		ArrayList<Objective> objectives = domain.getObjectives();
 		HashSet<String> names = new HashSet<String>();
-		
+
 		for (int i = 0; i < objectives.size(); i++) {
 			// Check indices
 			if (objectives.get(i).getNumber() != i) {
-				errors += ("Index of \"" + objectives.get(i).getName() + "\" is " + objectives.get(i).getNumber() + " but should be " + i + "\n");
+				errors += ("Index of \"" + objectives.get(i).getName()
+						+ "\" is " + objectives.get(i).getNumber()
+						+ " but should be " + i + "\n");
 			}
 			// Check unique names
 			if (!names.add(objectives.get(i).getName())) {
-				errors += ("There already exists an element with objective name \"" + objectives.get(i).getName() + "\"\n");
+				errors += ("There already exists an element with objective name \""
+						+ objectives.get(i).getName() + "\"\n");
 			}
 		}
 		return errors;
 	}
-	
-	private static String validateCorrespondenceDomainAndProfile(Domain domain, UtilitySpace space) {
-		
+
+	private static String validateCorrespondenceDomainAndProfile(Domain domain,
+			UtilitySpace space) {
+
 		String errors = isComplete(space);
 		if (errors == null) {
 			errors = "";
@@ -50,47 +56,55 @@ public class ScenarioValidator {
 		}
 		return errors;
 	}
-	
-    private static String isComplete(UtilitySpace space) 
+
+	private static String isComplete(UtilitySpace space)
 	// Oh damn, problem, we don't have the domain template here anymore.
-    // so how can we check domain compativility?
-    // only we can check that all fields are filled.........
-    { 
-    	Domain domain = space.getDomain();
-    	ArrayList<Issue> issues=domain.getIssues();
-    	if (issues==null) return "Utility space is not complete, in fact it is empty!";
-    	String mess;
-    	for (Issue issue:issues) 
-    	{
-    		Evaluator ev=space.getEvaluator(issue.getNumber());
-    		if (ev==null) return "issue "+issue.getName()+" has no evaluator";
-    		mess= (ev.isComplete(issue));
-    		if (mess!=null) return mess;
-    	}
-    	return null;
-    }
-	
+	// so how can we check domain compativility?
+	// only we can check that all fields are filled.........
+	{
+		Domain domain = space.getDomain();
+		ArrayList<Issue> issues = domain.getIssues();
+		if (issues == null)
+			return "Utility space is not complete, in fact it is empty!";
+		String mess;
+		for (Issue issue : issues) {
+			Evaluator ev = space.getEvaluator(issue.getNumber());
+			if (ev == null)
+				return "issue " + issue.getName() + " has no evaluator";
+			mess = (ev.isComplete(issue));
+			if (mess != null)
+				return mess;
+		}
+		return null;
+	}
+
 	private static String validatePreferenceProfile(UtilitySpace space) {
 		String errors = "";
 		double weightSum = 0;
 		for (Entry<Objective, Evaluator> pair : space.getEvaluators()) {
 			Objective obj = pair.getKey();
 			Evaluator eval = pair.getValue();
-			
+
 			if (obj == null) {
 				errors += "Mismatch in objective indices between domain and preference profile\n";
 				return errors;
 			}
-			
+
 			if (obj.getType() == ISSUETYPE.DISCRETE) {
 				EvaluatorDiscrete dEval = (EvaluatorDiscrete) eval;
 				boolean allZero = true;
 				for (ValueDiscrete dValue : dEval.getValues()) {
 					try {
-						int evaluation = dEval.getEvaluationNotNormalized(dValue);
-						
+						double evaluation = dEval
+								.getEvaluationNotNormalized(dValue);
+
 						if (evaluation < 0) {
-							errors += "Value \"" + dValue.getValue() + "\"" + " of objective \"" + obj.getName() + "\" must have a non-negative evaluation\n";
+							errors += "Value \""
+									+ dValue.getValue()
+									+ "\""
+									+ " of objective \""
+									+ obj.getName()
+									+ "\" must have a non-negative evaluation\n";
 						} else if (evaluation > 0) {
 							allZero = false;
 						}
@@ -99,7 +113,9 @@ public class ScenarioValidator {
 					}
 				}
 				if (allZero) {
-					errors += "Objective \"" + obj.getName() + "\" does not have a value with a non-zero positive evaluation\n";
+					errors += "Objective \""
+							+ obj.getName()
+							+ "\" does not have a value with a non-zero positive evaluation\n";
 				}
 			}
 			weightSum += eval.getWeight();
@@ -109,23 +125,27 @@ public class ScenarioValidator {
 		}
 		return errors;
 	}
-	
+
 	public static String validateDomainRepository(Repository domainrepository) {
 		String errors = "";
 		try {
 			for (RepItem repitem : domainrepository.getItems()) {
-				DomainRepItem dri = (DomainRepItem)repitem;
+				DomainRepItem dri = (DomainRepItem) repitem;
 				Domain domain = new Domain(dri.getURL().getFile());
-				
+
 				for (ProfileRepItem pri : dri.getProfiles()) {
-					UtilitySpace space = new UtilitySpace(domain, pri.getURL().getFile());
+					UtilitySpace space = new UtilitySpace(domain, pri.getURL()
+							.getFile());
 					String newErrors = "";
 					newErrors += (validateDomain(domain));
-					newErrors += (validateCorrespondenceDomainAndProfile(domain, space));
+					newErrors += (validateCorrespondenceDomainAndProfile(
+							domain, space));
 					newErrors += (validatePreferenceProfile(space));
 
 					if (!newErrors.trim().equals("")) {
-						errors += "DOMAIN: " + dri.getName() + " PROFILE: " + pri.getURL().getFile() + "\n" + newErrors + "\n";
+						errors += "DOMAIN: " + dri.getName() + " PROFILE: "
+								+ pri.getURL().getFile() + "\n" + newErrors
+								+ "\n";
 					}
 				}
 			}
@@ -134,7 +154,7 @@ public class ScenarioValidator {
 		}
 		return errors;
 	}
-	
+
 	public static void main(String[] args) {
 		try {
 			Repository domainrepository = Repository.get_domain_repos();
